Implemented MatchPlayActivity.
