+++ app/src/main/AndroidManifest.xml
@@ -33,7 +33,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".MatchPlayActivity"></activity>
+        <activity android:name=".match.MatchPlayActivity"></activity>
     </application>
 
 </manifest>
