+++ app/src/main/res/layout/activity_match_play.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="nl.jft.MatchPlayActivity">
+    tools:context="nl.jft.match.MatchPlayActivity">
 
     <Button
         android:id="@+id/match_play_button_player_two"
